chore: update dependencies while preserving entrypoint
