+++ package.json
@@ -6,7 +6,7 @@
   "type": "module",
   "main": "dist/index.js",
   "bin": {
-    "mcp-image": "dist/cli.js"
+    "mcp-image": "dist/index.js"
   },
   "keywords": [
     "mcp",
@@ -60,20 +60,19 @@
     "check:all": "npm run check && npm run check:deps && npm run check:unused && npm run build && npm run test"
   },
   "dependencies": {
-    "@google/genai": "^2.15.0",
+    "@google/genai": "^2.19.0",
     "@modelcontextprotocol/sdk": "^1.30.0",
-    "openai": "^7.3.0"
+    "openai": "^7.8.0"
   },
   "devDependencies": {
-    "@biomejs/biome": "^2.5.6",
-    "@types/node": "^26.1.2",
-    "@vitest/ui": "^4.1.10",
+    "@biomejs/biome": "^2.5.11",
+    "@types/node": "^26.4.0",
     "dpdm": "^4.3.0",
     "husky": "^9.1.7",
-    "knip": "^6.31.0",
-    "lint-staged": "^17.3.0",
+    "knip": "^6.33.0",
+    "lint-staged": "^17.4.1",
     "typescript": "^7.0.2",
-    "vitest": "^4.1.10"
+    "vitest": "^4.1.11"
   },
   "engines": {
     "node": ">=22"
