+++ src/__tests__/index.test.ts
@@ -1,18 +1,26 @@
-import { describe, expect, it, vi } from 'vitest'
-
-const serverEntry = vi.hoisted(() => ({ loaded: false }))
-
-vi.mock('../server-main.js', () => {
-  serverEntry.loaded = true
-  return {}
-})
+import { fileURLToPath } from 'node:url'
+import { Client } from '@modelcontextprotocol/sdk/client/index.js'
+import { StdioClientTransport } from '@modelcontextprotocol/sdk/client/stdio.js'
+import { describe, expect, it } from 'vitest'
 
 describe('package entry point', () => {
-  it('exports the library API without loading the stdio server entry point', async () => {
-    const api = await import('../index.js')
+  it('serves MCP over stdio when dist/index.js is executed directly', async () => {
+    const projectRoot = fileURLToPath(new URL('../..', import.meta.url))
+    const transport = new StdioClientTransport({
+      command: process.execPath,
+      args: ['dist/index.js'],
+      cwd: projectRoot,
+      stderr: 'pipe',
+    })
+    const client = new Client({ name: 'entry-point-test', version: '1.0.0' })
 
-    expect(api.createMCPServer).toBeTypeOf('function')
-    expect(api.MCPServerImpl).toBeTypeOf('function')
-    expect(serverEntry.loaded).toBe(false)
+    try {
+      await client.connect(transport)
+      const { tools } = await client.listTools()
+
+      expect(tools.map(({ name }) => name)).toEqual(['generate_image'])
+    } finally {
+      await client.close()
+    }
   })
 })
